#1405 Introduce LocalResult factory: simple refactoring
