+++ h2/src/main/org/h2/command/dml/Explain.java
@@ -16,6 +16,7 @@ import org.h2.expression.Expression;
 import org.h2.expression.ExpressionColumn;
 import org.h2.mvstore.db.MVTableEngine.Store;
 import org.h2.result.LocalResult;
+import org.h2.result.LocalResultFactory;
 import org.h2.result.ResultInterface;
 import org.h2.store.PageStore;
 import org.h2.table.Column;
@@ -72,7 +73,7 @@ public class Explain extends Prepared {
         Database db = session.getDatabase();
         ExpressionColumn expr = new ExpressionColumn(db, column);
         Expression[] expressions = { expr };
-        result = new LocalResult(session, expressions, 1);
+        result = LocalResultFactory.createRow(session, expressions, 1);
         if (maxrows >= 0) {
             String plan;
             if (executeCommand) {
