+++ h2/src/main/org/h2/engine/Database.java
@@ -41,6 +41,8 @@ import org.h2.message.Trace;
 import org.h2.message.TraceSystem;
 import org.h2.mvstore.MVStore;
 import org.h2.mvstore.db.MVTableEngine;
+import org.h2.result.LocalResult;
+import org.h2.result.LocalResultFactory;
 import org.h2.result.Row;
 import org.h2.result.RowFactory;
 import org.h2.result.SearchRow;
@@ -230,6 +232,7 @@ public class Database implements DataHandler {
     private int queryStatisticsMaxEntries = Constants.QUERY_STATISTICS_MAX_ENTRIES;
     private QueryStatisticsData queryStatisticsData;
     private RowFactory rowFactory = RowFactory.DEFAULT;
+    private LocalResultFactory resultFactory = LocalResultFactory.DEFAULT;
 
     private Authenticator authenticator;
 
@@ -367,6 +370,14 @@ public class Database implements DataHandler {
 
     public void setRowFactory(RowFactory rowFactory) {
         this.rowFactory = rowFactory;
+    }
+
+    public LocalResultFactory getLocalResultFactory() {
+        return resultFactory;
+    }
+
+    public void setLocalResultFactory(LocalResultFactory resultFactory) {
+        this.resultFactory = resultFactory;
     }
 
     public static void setInitialPowerOffCount(int count) {
