+++ h2/src/main/org/h2/engine/GeneratedKeys.java
@@ -13,6 +13,7 @@ import java.util.Map;
 import org.h2.expression.Expression;
 import org.h2.expression.ExpressionColumn;
 import org.h2.result.LocalResult;
+import org.h2.result.LocalResultFactory;
 import org.h2.result.Row;
 import org.h2.table.Column;
 import org.h2.table.Table;
@@ -128,7 +129,7 @@ public final class GeneratedKeys {
         Database db = session == null ? null : session.getDatabase();
         if (Boolean.FALSE.equals(generatedKeysRequest)) {
             clear(null);
-            return new LocalResult();
+            return LocalResultFactory.createRow(session);
         }
         ArrayList<ExpressionColumn> expressionColumns;
         if (Boolean.TRUE.equals(generatedKeysRequest)) {
@@ -152,7 +153,7 @@ public final class GeneratedKeys {
                 }
             } else {
                 clear(null);
-                return new LocalResult();
+                return LocalResultFactory.createRow(session);
             }
         } else if (generatedKeysRequest instanceof String[]) {
             if (table != null) {
@@ -182,18 +183,18 @@ public final class GeneratedKeys {
                 }
             } else {
                 clear(null);
-                return new LocalResult();
+                return LocalResultFactory.createRow(session);
             }
         } else {
             clear(null);
-            return new LocalResult();
+            return LocalResultFactory.createRow(session);
         }
         int columnCount = expressionColumns.size();
         if (columnCount == 0) {
             clear(null);
-            return new LocalResult();
+            return LocalResultFactory.createRow(session);
         }
-        LocalResult result = new LocalResult(session, expressionColumns.toArray(new Expression[0]), columnCount);
+        LocalResult result = LocalResultFactory.createRow(session, expressionColumns.toArray(new Expression[0]), columnCount);
         for (Map<Column, Value> map : data) {
             Value[] row = new Value[columnCount];
             for (Map.Entry<Column, Value> entry : map.entrySet()) {
